+++ server/src/internalClusterTest/java/org/opensearch/remotestore/RemoteRestoreSnapshotIT.java
@@ -740,8 +740,8 @@ public class RemoteRestoreSnapshotIT extends RemoteSnapshotIT {
         );
         assertTrue(exception.getMessage().contains("cannot modify setting [index.remote_store.segment.repository]" + " on restore"));
 
-        // try index restore with index.knn setting modified
-        Settings indexKnnSettings = Settings.builder().put(IndexMetadata.SETTING_KNN_ENABLED, false).build();
+        // try index restore with index.number_of_shards setting modified
+        Settings indexKnnSettings = Settings.builder().put(IndexMetadata.SETTING_NUMBER_OF_SHARDS, false).build();
 
         exception = expectThrows(
             SnapshotRestoreException.class,
@@ -755,7 +755,7 @@ public class RemoteRestoreSnapshotIT extends RemoteSnapshotIT {
                 .setRenameReplacement(restoredIndex)
                 .get()
         );
-        assertTrue(exception.getMessage().contains("cannot modify setting [index.knn]" + " on restore"));
+        assertTrue(exception.getMessage().contains("cannot modify UnmodifiableOnRestore setting [index.number_of_shards]" + " on restore"));
 
         // try index restore with remote store repository and translog store repository disabled
         exception = expectThrows(