+++ server/src/main/java/org/opensearch/snapshots/RestoreService.java
@@ -77,6 +77,7 @@ import org.opensearch.common.UUIDs;
 import org.opensearch.common.lucene.Lucene;
 import org.opensearch.common.regex.Regex;
 import org.opensearch.common.settings.ClusterSettings;
+import org.opensearch.common.settings.IndexScopedSettings;
 import org.opensearch.common.settings.Settings;
 import org.opensearch.common.unit.TimeValue;
 import org.opensearch.common.util.ArrayUtils;
@@ -120,16 +121,15 @@ import static org.opensearch.cluster.metadata.IndexMetadata.SETTING_AUTO_EXPAND_
 import static org.opensearch.cluster.metadata.IndexMetadata.SETTING_CREATION_DATE;
 import static org.opensearch.cluster.metadata.IndexMetadata.SETTING_HISTORY_UUID;
 import static org.opensearch.cluster.metadata.IndexMetadata.SETTING_INDEX_UUID;
-import static org.opensearch.cluster.metadata.IndexMetadata.SETTING_KNN_ENABLED;
 import static org.opensearch.cluster.metadata.IndexMetadata.SETTING_NUMBER_OF_REPLICAS;
 import static org.opensearch.cluster.metadata.IndexMetadata.SETTING_NUMBER_OF_SEARCH_REPLICAS;
-import static org.opensearch.cluster.metadata.IndexMetadata.SETTING_NUMBER_OF_SHARDS;
 import static org.opensearch.cluster.metadata.IndexMetadata.SETTING_REMOTE_SEGMENT_STORE_REPOSITORY;
 import static org.opensearch.cluster.metadata.IndexMetadata.SETTING_REMOTE_STORE_ENABLED;
 import static org.opensearch.cluster.metadata.IndexMetadata.SETTING_REMOTE_TRANSLOG_STORE_REPOSITORY;
 import static org.opensearch.cluster.metadata.IndexMetadata.SETTING_REPLICATION_TYPE;
 import static org.opensearch.cluster.metadata.IndexMetadata.SETTING_VERSION_CREATED;
 import static org.opensearch.cluster.metadata.IndexMetadata.SETTING_VERSION_UPGRADED;
+import static org.opensearch.cluster.metadata.IndexMetadata.SETTING_NUMBER_OF_SHARDS;
 import static org.opensearch.common.util.FeatureFlags.SEARCHABLE_SNAPSHOT_EXTENDED_COMPATIBILITY;
 import static org.opensearch.common.util.IndexUtils.filterIndices;
 import static org.opensearch.common.util.set.Sets.newHashSet;
@@ -165,15 +165,13 @@ public class RestoreService implements ClusterStateApplier {
 
     private static final Set<String> USER_UNMODIFIABLE_SETTINGS = unmodifiableSet(
         newHashSet(
-            SETTING_NUMBER_OF_SHARDS,
             SETTING_VERSION_CREATED,
             SETTING_INDEX_UUID,
             SETTING_CREATION_DATE,
             SETTING_HISTORY_UUID,
             SETTING_REMOTE_STORE_ENABLED,
             SETTING_REMOTE_SEGMENT_STORE_REPOSITORY,
-            SETTING_REMOTE_TRANSLOG_STORE_REPOSITORY,
-            SETTING_KNN_ENABLED
+            SETTING_REMOTE_TRANSLOG_STORE_REPOSITORY
         )
     );
 
@@ -184,6 +182,7 @@ public class RestoreService implements ClusterStateApplier {
     static {
         Set<String> unremovable = new HashSet<>(USER_UNMODIFIABLE_SETTINGS.size() + 4);
         unremovable.addAll(USER_UNMODIFIABLE_SETTINGS);
+        unremovable.add(SETTING_NUMBER_OF_SHARDS);
         unremovable.add(SETTING_NUMBER_OF_REPLICAS);
         unremovable.add(SETTING_AUTO_EXPAND_REPLICAS);
         unremovable.add(SETTING_VERSION_UPGRADED);
@@ -203,6 +202,8 @@ public class RestoreService implements ClusterStateApplier {
     private final ShardLimitValidator shardLimitValidator;
 
     private final ClusterSettings clusterSettings;
+
+    private final IndexScopedSettings indexScopedSettings;
 
     private final IndicesService indicesService;
 
@@ -236,6 +237,7 @@ public class RestoreService implements ClusterStateApplier {
         this.clusterSettings = clusterService.getClusterSettings();
         this.shardLimitValidator = shardLimitValidator;
         this.indicesService = indicesService;
+        this.indexScopedSettings = indicesService.getIndexScopedSettings();
         this.clusterInfoSupplier = clusterInfoSupplier;
         this.dataToFileCacheSizeRatioSupplier = dataToFileCacheSizeRatioSupplier;
 
@@ -874,6 +876,11 @@ public class RestoreService implements ClusterStateApplier {
                             .put(normalizedChangeSettings.filter(k -> {
                                 if (USER_UNMODIFIABLE_SETTINGS.contains(k)) {
                                     throw new SnapshotRestoreException(snapshot, "cannot modify setting [" + k + "] on restore");
+                                } else if (indexScopedSettings.isUnmodifiableOnRestoreSetting(k)) {
+                                    throw new SnapshotRestoreException(
+                                        snapshot,
+                                        "cannot modify UnmodifiableOnRestore setting [" + k + "] on restore"
+                                    );
                                 } else {
                                     return true;
                                 }